Rename hpe_morpheus_workflow_catalog_item to hpe_morpheus_catalog_item_workflow
