+++ examples/resources/morpheus_catalog_item_workflow/import.sh
@@ -1,0 +1,1 @@
+terraform import hpe_morpheus_catalog_item_workflow.tf_example_workflow_catalog_item 1
