--- examples/resources/morpheus_workflow_catalog_item/import.sh
@@ -1,1 +1,0 @@
-terraform import hpe_morpheus_workflow_catalog_item.tf_example_workflow_catalog_item 1
