+++ templates/resources/morpheus_catalog_item_workflow.md.tmpl
@@ -10,7 +10,7 @@ description: |-
 
 ## Example Usage
 
-{{tffile "examples/resources/morpheus_workflow_catalog_item/resource.tf"}}
+{{tffile "examples/resources/morpheus_catalog_item_workflow/resource.tf"}}
 
 {{ .SchemaMarkdown | trimspace }}
 
@@ -18,4 +18,4 @@ description: |-
 
 Import is supported using the following syntax:
 
-{{codefile "shell" "examples/resources/morpheus_workflow_catalog_item/import.sh" }}
+{{codefile "shell" "examples/resources/morpheus_catalog_item_workflow/import.sh" }}
